# Optimize BinaryHeap bounds checking