@Aatch 2016-08-28
Could you add some comments explaining why the code is safe. As in, why we can assume the index is valid? The comments on the functions themselves should probably reflect the new preconditions and be updated since they won't panic in those cases now.