@arthurprs 2016-08-29
Updated the PR with @Aatch suggestions.

I also took the liberty of annotating the rest of Hole<> functions with inline and removing the always, this looks cleaner and more consistent.