@alexcrichton 2016-09-02
@bors: retry

On Fri, Sep 2, 2016 at 8:02 AM, bors <notifications@github.com> wrote:

> 💔 Test failed - auto-win-gnu-32-opt-rustbuild
> <https://buildbot.rust-lang.org/builders/auto-win-gnu-32-opt-rustbuild/builds/2375>
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/rust-lang/rust/pull/36072#issuecomment-244400250>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAD95Epip6mm_z-13xfqsm9GRh1RQhLlks5qmDqOgaJpZM4Ju75f>
> .
>