@alexcrichton 2016-09-02
@bors: retry