@alexcrichton 2016-09-02
@bors: Retry

On Fri, Sep 2, 2016 at 9:13 AM, bors <notifications@github.com> wrote:

> 💔 Test failed - auto-win-msvc-64-cargotest
> <https://buildbot.rust-lang.org/builders/auto-win-msvc-64-cargotest/builds/1712>
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/rust-lang/rust/pull/36072#issuecomment-244419386>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAD95NSdzh7XzwhIL08BQKyUdH-jSpskks5qmEsOgaJpZM4Ju75f>
> .
>